Some preliminary work on refactoring
